fix: move revert functions to pathfinder::state::sync
