+++ crates/merkle-tree/src/class.rs
@@ -2,8 +2,7 @@ use std::collections::HashMap;
 
 use anyhow::Context;
 use pathfinder_common::{
-    calculate_class_commitment_leaf_hash, BlockNumber, ClassCommitment, ClassCommitmentLeafHash,
-    ClassHash, SierraHash,
+    BlockNumber, ClassCommitment, ClassCommitmentLeafHash, ClassHash, SierraHash,
 };
 use pathfinder_crypto::Felt;
 use pathfinder_storage::{Node, Transaction};
@@ -113,71 +112,4 @@ impl crate::storage::Storage for ClassStorage<'_> {
 
         Ok(leaf)
     }
-}
-
-/// Revert all class trie updates.
-pub fn revert_class_updates(
-    transaction: &Transaction<'_>,
-    head: BlockNumber,
-    target_block: BlockNumber,
-    expected_class_commitment: ClassCommitment,
-    force: bool,
-) -> anyhow::Result<()> {
-    let class_root_exists = match transaction.class_root_index(target_block)? {
-        None => false,
-        Some(root_idx) => transaction.class_trie_node(root_idx)?.is_some(),
-    };
-
-    if force || !class_root_exists {
-        let updates = transaction.reverse_sierra_class_updates(head, target_block)?;
-
-        let mut class_tree = ClassCommitmentTree::load(transaction, head)
-            .context("Loading class commitment trie")?;
-
-        for (class_hash, casm_update) in updates {
-            let new_value = match casm_update {
-                None => {
-                    // The class must be removed
-                    ClassCommitmentLeafHash::ZERO
-                }
-                Some(casm_hash) => {
-                    // Class hash has changed. Note that the class commitment leaf must have already been added to storage.
-                    calculate_class_commitment_leaf_hash(casm_hash)
-                }
-            };
-
-            class_tree
-                .set(class_hash, new_value)
-                .context("Updating class commitment trie")?;
-        }
-
-        let (class_commitment, nodes_added) =
-            class_tree.commit().context("Committing class trie")?;
-
-        if expected_class_commitment != class_commitment {
-            anyhow::bail!(
-                "Storage commitment mismatch: expected {}, calculated {}",
-                expected_class_commitment,
-                class_commitment
-            );
-        }
-
-        let root_idx = if !class_commitment.0.is_zero() {
-            let root_idx = transaction
-                .insert_class_trie(class_commitment, &nodes_added)
-                .context("Persisting class trie")?;
-
-            Some(root_idx)
-        } else {
-            None
-        };
-
-        transaction
-            .insert_or_update_class_root(target_block, root_idx)
-            .context("Inserting class root index")?;
-
-        tracing::debug!(%target_block, %class_commitment, "Committed class trie");
-    }
-
-    Ok(())
 }
