+++ crates/pathfinder/src/state.rs
@@ -1,4 +1,4 @@
 pub mod block_hash;
 mod sync;
 
-pub use sync::{l1, l2, sync, Gossiper, SyncContext};
+pub use sync::{l1, l2, revert, sync, Gossiper, SyncContext};
